Update main.py
Added single_sided_cards and double_sided_cards functions for readability.

Added a context manager to remove the need for the close() statement

Changed the try-except KeyError on the bottom to an if statement instead.

Moved the response variable and the JS variable to remove repeating code.

Updated the url to an f-string.